refactor(protocol): remove proven T1 debris
