--- packages/protocol/src/capabilities/contacts.tools.port.ts
@@ -1,9 +1,0 @@
-/**
- * @deprecated — compatibility shim.
- *
- * Canonical type has moved to contacts/ports.
- * This re-export exists solely to avoid breaking imports that pre-date IND-549.
- *
- * IND-549: migrated to src/contacts/ports/contact.tools.port.ts.
- */
-export type { ContactToolDeps } from '../contacts/ports/index.js';
