--- packages/protocol/src/capabilities/integrations.tools.port.ts
@@ -1,9 +1,0 @@
-/**
- * @deprecated — compatibility shim.
- *
- * Canonical type has moved to integrations/ports.
- * This re-export exists solely to avoid breaking imports that pre-date IND-549.
- *
- * IND-549: migrated to src/integrations/ports/integration.tools.port.ts.
- */
-export type { IntegrationToolDeps } from '../integrations/ports/index.js';
