--- packages/protocol/src/capabilities/negotiation.tools.port.ts
@@ -1,5 +1,0 @@
-/**
- * Thin backward-compat shim for the negotiation tool-host port — IND-550.
- * Canonical location: negotiation/ports/negotiation.tools.port.ts
- */
-export type { NegotiationToolDeps } from "../negotiation/ports/negotiation.tools.port.js";
