--- packages/protocol/src/capabilities/opportunities.tools.port.ts
@@ -1,5 +1,0 @@
-/**
- * Thin backward-compat shim — IND-551.
- * Canonical location: opportunity/ports/opportunity.tools.port.ts
- */
-export type { OpportunityToolDeps } from "../opportunity/ports/index.js";
