--- packages/protocol/src/capabilities/participant-agents.tools.port.ts
@@ -1,9 +1,0 @@
-/**
- * @deprecated — compatibility shim.
- *
- * Canonical type has moved to participant-agents/ports.
- * This re-export exists solely to avoid breaking imports that pre-date IND-548.
- *
- * IND-548: migrated to src/participant-agents/ports/agent.tools.port.ts.
- */
-export type { AgentToolDeps } from '../participant-agents/ports/index.js';
