--- packages/protocol/src/capabilities/questions.tools.port.ts
@@ -1,5 +1,0 @@
-/**
- * @deprecated Canonical location: questions/ports/question.tools.port — retained
- * for backward compatibility (IND-547).
- */
-export type { QuestionerToolDeps, AskUserQuestionToolDeps } from "../questions/ports/question.tools.port.js";
