+++ packages/protocol/src/contacts/index.ts
@@ -6,19 +6,14 @@
  * contacts/ports; this barrel is for cross-capability consumers that must
  * go through the contacts public surface.
  *
- * IND-549: canonical home for contact management capabilities previously
- * spread across contact/, shared/interfaces/contact.interface.ts, and
- * capabilities/contacts.tools.port.ts.
+ * IND-549: canonical home for contact management capabilities.
  *
  * Retains participant reachability semantics: all operations are scoped to
  * the owning user's personal network. Contacts are participants with whom
  * opportunity discovery is run.
  *
- * Legacy paths:
- * - contact/contact.tools.ts — thin compatibility shim → contacts/application
- * - contact/contact.inviter.ts — thin compatibility shim → contacts/application
+ * Compatibility paths:
  * - shared/interfaces/contact.interface.ts — thin compatibility shim → contacts/ports
- * - capabilities/contacts.tools.port.ts — thin compatibility shim → contacts/ports
  * - capabilities/contacts.facade.ts — thin re-export via contacts/public
  */
 export * from "./public/index.js";
